Feat: writer 및 과신고 검증 메서드를 Guestbook에서 Postings로 옮김
